Finished all ports

+++ io.fairyproject.modules/platform.mc/module.protocol/mc-protocol/src/main/java/io/fairyproject/mc/protocol/packet/client/CPacketWindowClick.java
@@ -9,7 +9,7 @@ public interface CPacketWindowClick extends CPacket {
 
     int getWindowButton();
 
-    Optional<Short> getActionNumber();
+    Optional<Integer> getActionNumber();
 
     int getMode();
